agent/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
google-cloud-aiplatform
vertexai
opencv-python-headless
pillow
